use moment.js, and normalized title
