+++ package.json
@@ -20,11 +20,13 @@
     "babel-preset-stage-0": "^6.24.1",
     "highlight.js": "^9.12.0",
     "lodash": "^4.17.10",
+    "moment": "^2.24.0",
     "react": "^16.3.2",
     "react-addons-shallow-compare": "^15.6.2",
     "react-dom": "^16.3.2",
     "react-lowlight": "^2.0.0",
     "react-markdown": "^2.4.0",
+    "react-moment": "^0.8.4",
     "react-redux": "^5.0.7",
     "react-router-dom": "^4.2.2",
     "react-scroll": "^1.7.9",
